Add files via upload
The train.csv and python file are added. please open it in spyder or jupyter notebook or other editors.
some classification such as Logistic regression, Knn, discriminant analysis are applied and there are some visualizationof features.
It is not complete and there are some other classification which should be applied to find the best classification and have accurate prediction.